Создает класс Store, реализует методы получения и записи данных в localStorage, подключает в main.js

+++ src/api/provider.js
@@ -1,21 +1,43 @@
+import TaskModel from "../models/task.js";
+
+const isOnline = () => window.navigator.onLine;
+
 export default class Provider {
-  constructor(api) {
+  constructor(api, store) {
     this._api = api;
+    this._store = store;
   }
 
   getTasks() {
-    return this._api.getTasks();
+    if (isOnline()) {
+      return this._api.getTasks()
+        .then((tasks) => {
+          tasks.forEach((task) => this._store.setItem(task.id, task.convertToRaw()));
+          return tasks;
+        });
+    }
+    const storedTasks = Object.values(this._store.getItems());
+    return Promise.resolve(TaskModel.parseTasks(storedTasks));
   }
 
   updateTasks(id, data) {
-    return this._api.updateTasks(id, data);
+    if (isOnline()) {
+      return this._api.updateTasks(id, data);
+    }
+    return Promise.reject(`Offline logic is not implemented`);
   }
 
   createTask(data) {
-    return this._api.createTask(data);
+    if (isOnline()) {
+      return this._api.createTask(data);
+    }
+    return Promise.reject(`Offline logic is not implemented`);
   }
 
   deleteTask(id) {
-    return this._api.deleteTask(id);
+    if (isOnline()) {
+      return this._api.deleteTask(id);
+    }
+    return Promise.reject(`Offline logic is not implemented`);
   }
 }
